step: employ disc vector for static image
step:
+ Add initial logic to method animate so that this method employs vector
    discs to render a static image of a disc.
+ In main, call methods setDiscs and animate from the Display object.

fix:
* In struct Circ, make attribute radius an unsigned short.
* Replace any int declaration in RenderFillCirc by one for a short.

chore:
- Remove commented-out code.

+++ src/fillCirc.cpp
@@ -3,13 +3,13 @@
 // using Circ, IntPair, SDL_Renderer and SDL_RenderDrawPoint
 
 void RenderFillCirc(SDL_Renderer* renderer, Circ* circ) {
-  // Comparison between signed int i and unsigned int u: i <= u.
-  auto leq = [](int i, unsigned int u)->bool {
-    return  i < 0  ||  static_cast<unsigned int>(i) <= u;
+  // Comparison between signed short i and unsigned short u: i <= u.
+  auto leq = [](short i, unsigned short u)->bool {
+    return  i < 0  ||  static_cast<unsigned short>(i) <= u;
   };
 
-  const unsigned int rad = circ->radius;
-  int px, py; // pixel coordinates relative to circle center
+  const unsigned short rad = circ->radius;
+  short px, py; // pixel coordinates relative to circle center
   for( px = -rad;  leq(px, rad);  ++px )
   for( py = -rad;  leq(py, rad);  ++py )
     if( leq(px*px + py*py,  rad*rad) )
